`constrained-generators`: Improve documentation in a bunch of places
