+++ libs/constrained-generators/examples/Constrained/Examples.hs
