+++ libs/constrained-generators/examples/Constrained/Examples/CheatSheet.hs
