+++ libs/constrained-generators/examples/Constrained/Examples/ManualExamples.hs
