+++ libs/constrained-generators/examples/Constrained/Examples/Tree.hs
@@ -4,7 +4,7 @@
 module Constrained.Examples.Tree where
 
 import Constrained.API
-import Constrained.Spec.Tree (BinTree (..))
+import Constrained.Examples.BinTree
 import Data.Tree
 
 allZeroTree :: Specification (BinTree Int)
