+++ libs/constrained-generators/src/Constrained/API/Extend.hs
@@ -1,5 +1,7 @@
 {-# LANGUAGE PatternSynonyms #-}
 
+-- | This module provides an API for extending the library with new function
+-- symbols.
 module Constrained.API.Extend (
   -- * Abstract syntax
   SpecificationD (..),
@@ -27,6 +29,22 @@ module Constrained.API.Extend (
   NumSpec (..),
   TreeSpec (..),
 
+  -- * Generics
+  (:::),
+  SOP,
+  algebra,
+  inject,
+
+  -- * Building new `NumSpec`-based instances
+  emptyNumSpec,
+  cardinalNumSpec,
+  combineNumSpec,
+  genFromNumSpec,
+  shrinkWithNumSpec,
+  conformsToNumSpec,
+  toPredsNumSpec,
+  MaybeBounded (..),
+
   -- * Re-export of `Constrained.API`
   module Constrained.API,
 ) where
@@ -35,8 +53,9 @@ import Constrained.API
 import Constrained.AbstractSyntax
 import Constrained.Base
 import Constrained.FunctionSymbol
+import Constrained.Generic
 import Constrained.NumOrd
 import Constrained.Spec.Map
 import Constrained.Spec.Set
-import Constrained.Spec.SumProd
 import Constrained.Spec.Tree
+import Constrained.TheKnot
