+++ libs/constrained-generators/src/Constrained/GenT.hs
@@ -3,12 +3,61 @@
 {-# LANGUAGE ImportQualifiedPost #-}
 {-# LANGUAGE RankNTypes #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TupleSections #-}
 {-# LANGUAGE TypeApplications #-}
 -- NOTE: this is for `split` vs. `splitGen` that we haven't had
 -- time to fix in `QuickCheck`.
 {-# OPTIONS_GHC -Wno-deprecations #-}
 
-module Constrained.GenT where
+-- | This module provides an interface for writing and working with generators
+--     that may fail in both recoverable and unrecoverable ways.
+module Constrained.GenT (
+  -- * Types
+  GE (..),
+  GenT,
+  GenMode (..),
+
+  -- * Writing t`GenT` generators
+  MonadGenError (..),
+  pureGen,
+  genFromGenT,
+  suchThatT,
+  suchThatWithTryT,
+  scaleT,
+  firstGenT,
+  tryGenT,
+  chooseT,
+  sizeT,
+  withMode,
+  frequencyT,
+  oneofT,
+  vectorOfT,
+  listOfUntilLenT,
+  listOfT,
+  resizeT,
+  strictGen,
+  looseGen,
+
+  -- * So far undocumented
+  fatalError,
+  catMessages,
+  catMessageList,
+  explain,
+  errorGE,
+  fromGE,
+  runGE,
+  inspect,
+  genError,
+  pushGE,
+  push,
+  dropGen,
+  catchGen,
+  getMode,
+  headGE,
+  fromGEProp,
+  fromGEDiscard,
+  listFromGE,
+) where
 
 import Control.Monad
 import Data.Foldable
@@ -23,8 +72,8 @@ import Test.QuickCheck.Gen
 -- ==============================================================
 -- The GE Monad
 
--- | It distinguishes between two kinds of errors: FatalError and GenError
--- and non-fatal errors.
+-- | This is like an @Error@ monad that distinguishes between two kinds of
+-- errors: @FatalError@s and non-fatal @GenError@s.
 data GE a
   = FatalError (NonEmpty (NonEmpty String))
   | GenError (NonEmpty (NonEmpty String))
@@ -56,6 +105,8 @@ data GenMode
   | Strict
   deriving (Ord, Eq, Show)
 
+-- | A `Gen` monad wrapper that allows different generation modes and different
+-- failure types.
 newtype GenT m a = GenT {runGenT :: GenMode -> [NonEmpty String] -> Gen (m a)}
   deriving (Functor)
 
@@ -78,8 +129,8 @@ instance MonadGenError m => MonadFail (GenT m) where
 ----------------------------------------------------------------------
 
 -- | A class for different types of errors with a stack of `explain` calls to
--- narrow down problems. The (NonEmpty String) means one cannot cause an
--- Error without at least 1 string to explain it.
+-- narrow down problems. The @NonEmpty String@ means one cannot cause an error
+-- without at least one string to explain it.
 class Monad m => MonadGenError m where
   genErrors :: HasCallStack => NonEmpty (NonEmpty String) -> m a
   fatalErrors :: HasCallStack => NonEmpty (NonEmpty String) -> m a
@@ -87,15 +138,15 @@ class Monad m => MonadGenError m where
   fatalErrorNE :: HasCallStack => NonEmpty String -> m a
   explainNE :: HasCallStack => NonEmpty String -> m a -> m a
 
--- | genError with one line of explanation
+-- | A potentially recoverable generation error
 genError :: MonadGenError m => String -> m a
 genError = genErrorNE . pure
 
--- | fatalError with one line of explanation
+-- | A non-recoverable fatal error
 fatalError :: MonadGenError m => String -> m a
 fatalError = fatalErrorNE . pure
 
--- | explain with one line of explanation
+-- | Attach an explanation to a computation in case of error
 explain :: MonadGenError m => String -> m a -> m a
 explain s
   | null s = id
@@ -137,13 +188,13 @@ cat :: NonEmpty (NonEmpty a) -> [NonEmpty a] -> NonEmpty (NonEmpty a)
 cat a [] = a
 cat a (x : xs) = a <> (x :| xs)
 
--- | Sometimes we have a bunch of genError or fatalError
---   messages we want to combine into one big message.
---   This happens when we want to lift one of these into an input for 'error'
+-- | Sometimes we have a bunch of `genError` or `fatalError` messages we want
+-- to combine into one big message. This happens when we want to lift one of
+-- these into an input for 'error'
 catMessages :: NonEmpty (NonEmpty String) -> String
 catMessages xs = unlines (NE.toList (catMessageList xs))
 
--- | Turn each inner (NonEmpty String) into a String
+-- | Turn each inner @NonEmpty String@ into a String
 catMessageList :: NonEmpty (NonEmpty String) -> NonEmpty String
 catMessageList = fmap (unlines . NE.toList)
 
@@ -151,8 +202,8 @@ catMessageList = fmap (unlines . NE.toList)
 -- Useful operations on GE
 
 -- If none of the GE's are FatalError, then concat together all the
--- Result's (skipping over GenError). If there is at least one
--- (FatalError xs) abort, and lift all those 'xs' as errors in the monad 'm'
+-- Results (skipping over GenError). If there is at least one
+-- @FatalError xs@ abort, and lift all those @xs@ as errors in the monad @m@.
 catGEs :: forall m a. MonadGenError m => [GE a] -> m [a]
 catGEs ges0 = go [] ges0
   where
@@ -163,22 +214,18 @@ catGEs ges0 = go [] ges0
         GenError _ -> go acc ges
         FatalError xs -> fatalErrors xs
 
--- | Given a function for handling GenError,
---   and handling FatalError by using 'error'
---   Turn (GE a) into 'a'
+-- | Turn @'GE' a@ into @a@ given a function for handling @GenError@, and handle
+-- @FatalError@ with 'error'
 fromGE :: HasCallStack => (NonEmpty (NonEmpty String) -> a) -> GE a -> a
 fromGE f ge = case ge of
   Result a -> a
   GenError xs -> f xs
   FatalError es -> error $ catMessages es
 
--- | Turn (GE a) into a
---   both GenError and FatalErrors are handled by using 'error'
+-- | Turn @'GE' a@ into where both @GenError@ and @FatalError@ are handled by
+-- using 'error'
 errorGE :: GE a -> a
-errorGE ge = case ge of
-  FatalError xs -> error $ catMessages xs
-  GenError xs -> error $ catMessages xs
-  Result x -> x
+errorGE = fromGE (error . catMessages)
 
 isOk :: GE a -> Bool
 isOk ge = case ge of
@@ -186,23 +233,29 @@ isOk ge = case ge of
   FatalError {} -> False
   Result {} -> True
 
+-- | Convert a `GE` into an arbitrary monad that has an instance of
+-- `MonadGenError`
 runGE :: forall m r. MonadGenError m => GE r -> m r
 runGE ge = case ge of
   GenError es -> genErrors es
   FatalError es -> fatalErrors es
   Result a -> pure a
 
+-- | Turn a `GE` for something testable into a `Property`, failing on any
+-- kind of error.
 fromGEProp :: Testable p => GE p -> Property
 fromGEProp ge = case ge of
   GenError es -> counterexample (catMessages es) False
   FatalError es -> counterexample (catMessages es) False
   Result p -> property p
 
+-- | Turn a `GE` into a property, `discard`ing any failure.
 fromGEDiscard :: Testable p => GE p -> Property
 fromGEDiscard ge = case ge of
   Result p -> property p
   _ -> discard
 
+-- | Like `Prelude.head` in the `GE` monad
 headGE :: Foldable t => t a -> GE a
 headGE t
   | x : _ <- toList t = pure x
@@ -215,28 +268,43 @@ listFromGE = fromGE (const []) . explain "listFromGE"
 -- ========================================================
 -- Useful operations on GenT
 
+-- | Run a t`GenT` generator in `Strict` mode
 strictGen :: GenT m a -> Gen (m a)
 strictGen genT = runGenT genT Strict []
 
-genFromGenT :: GenT GE a -> Gen a
-genFromGenT genT = errorGE <$> runGenT genT Strict []
+-- | Run a t`GenT` generator in `Loose` mode
+looseGen :: GenT m a -> Gen (m a)
+looseGen genT = runGenT genT Loose []
 
+-- | Turn a t`GenT` generator into a `Gen` generator in `Strict` mode
+genFromGenT :: GenT GE a -> Gen a
+genFromGenT genT = errorGE <$> strictGen genT
+
+-- | Locally change the generation size
 resizeT :: (Int -> Int) -> GenT m a -> GenT m a
 resizeT f (GenT gm) = GenT $ \mode msgs -> sized $ \sz -> resize (f sz) (gm mode msgs)
 
+-- | Turn a `Gen` generator into a t`GenT` generator that never fails.
 pureGen :: Applicative m => Gen a -> GenT m a
 pureGen gen = GenT $ \_ _ -> pure <$> gen
 
+-- | Lift `listOf` to t`GenT`
 listOfT :: MonadGenError m => GenT GE a -> GenT m [a]
 listOfT gen = do
   lst <- pureGen . listOf $ runGenT gen Loose []
   catGEs lst
 
--- | Generate a list of elements of length at most `goalLen`, but accepting failure
--- to get that many elements so long as `validLen` is true.
--- TODO: possibly one could return "more, fewer, ok" in the `validLen` instead
--- of `Bool`
-listOfUntilLenT :: (Typeable a, MonadGenError m) => GenT GE a -> Int -> (Int -> Bool) -> GenT m [a]
+-- | Generate a list of elements of length at most @goalLen@, but accepting
+-- failure to get that many elements so long as @validLen@ is true.
+listOfUntilLenT ::
+  (Typeable a, MonadGenError m) =>
+  -- | Element generator
+  GenT GE a ->
+  -- | @goalLen@ goal length
+  Int ->
+  -- | @validLen@ filter
+  (Int -> Bool) ->
+  GenT m [a]
 listOfUntilLenT gen goalLen validLen =
   genList `suchThatT` validLen . length
   where
@@ -244,6 +312,7 @@ listOfUntilLenT gen goalLen validLen =
       res <- pureGen . vectorOf goalLen $ runGenT gen Loose []
       catGEs res
 
+-- | Lift `vectorOf` to t`GenT`
 vectorOfT :: MonadGenError m => Int -> GenT GE a -> GenT m [a]
 vectorOfT i gen = GenT $ \mode _ -> do
   res <- fmap sequence . vectorOf i $ runGenT gen Strict []
@@ -255,9 +324,14 @@ vectorOfT i gen = GenT $ \mode _ -> do
 
 infixl 2 `suchThatT`
 
+-- | Lift `suchThat` to t`GenT`, equivalent to @`suchThatT` 100@
 suchThatT :: (Typeable a, MonadGenError m) => GenT m a -> (a -> Bool) -> GenT m a
 suchThatT g p = suchThatWithTryT 100 g p
 
+-- | Lift `suchThat` to t`GenT` with special handling of generation mode. In
+-- `Strict` mode @suchThatWithTry tries@ will try @tries@ times and fail with a
+-- `fatalError` if unsuccessful.  In `Loose` mode however, we will try only
+-- once and generate a `genError`.
 suchThatWithTryT ::
   forall a m. (Typeable a, MonadGenError m) => Int -> GenT m a -> (a -> Bool) -> GenT m a
 suchThatWithTryT tries g p = do
@@ -274,27 +348,27 @@ suchThatWithTryT tries g p = do
       a <- g
       if p a then pure a else scaleT (+ 1) $ go (n - 1) cont
 
+-- | Lift `scale` to t`GenT`
 scaleT :: (Int -> Int) -> GenT m a -> GenT m a
 scaleT sc (GenT gen) = GenT $ \mode msgs -> scale sc $ gen mode msgs
 
+-- | Access the `GenMode` we are running in, useful to decide e.g.  if we want
+-- to re-try in case of a `GenError` or give up
 getMode :: Applicative m => GenT m GenMode
 getMode = GenT $ \mode _ -> pure (pure mode)
 
 getMessages :: Applicative m => GenT m [NonEmpty String]
 getMessages = GenT $ \_ msgs -> pure (pure msgs)
 
+-- | Locally change the generation mode
 withMode :: GenMode -> GenT m a -> GenT m a
 withMode mode gen = GenT $ \_ msgs -> runGenT gen mode msgs
 
+-- | Lift `oneof` to t`GenT`
 oneofT :: (Typeable a, MonadGenError m) => [GenT GE a] -> GenT m a
-oneofT gs = do
-  mode <- getMode
-  msgs <- getMessages
-  r <-
-    explain "suchThatT in oneofT" $
-      pureGen (oneof [runGenT g mode msgs | g <- gs]) `suchThatT` isOk
-  runGE r
+oneofT gs = frequencyT $ map (1,) gs
 
+-- | Lift `frequency` to t`GenT`
 frequencyT :: (Typeable a, MonadGenError m) => [(Int, GenT GE a)] -> GenT m a
 frequencyT gs = do
   mode <- getMode
@@ -304,11 +378,13 @@ frequencyT gs = do
       pureGen (frequency [(f, runGenT g mode msgs) | (f, g) <- gs]) `suchThatT` isOk
   runGE r
 
+-- | Lift `choose` to t`GenT`, failing with a `genError` in case of an empty interval
 chooseT :: (Random a, Ord a, Show a, MonadGenError m) => (a, a) -> GenT m a
 chooseT (a, b)
   | b < a = genError ("chooseT (" ++ show a ++ ", " ++ show b ++ ")")
   | otherwise = pureGen $ choose (a, b)
 
+-- | Get the size provided to the generator
 sizeT :: Monad m => GenT m Int
 sizeT = GenT $ \mode msgs -> sized $ \n -> runGenT (pure n) mode msgs
 
@@ -360,12 +436,10 @@ firstGenT gs = loop gs []
         Left zs -> loop xs (zs : ys)
         Right a -> pure (Right a)
 
-liftGen :: forall x. (forall m. MonadGenError m => GenT m x) -> GenT GE x
-liftGen x = x
-
--- Drop a (GenT GE) computation into a (GenT m) computation.
--- Depending on the monad 'm' Some error information might be lost as
--- the monad m might fold FatalError's and GenError's together.
+-- | Drop a @t`GenT` `GE`@ computation into a @t`GenT` m@ computation.
+--
+-- Depending on the monad @m@ Some error information might be lost as
+-- the monad might fold `FatalError`'s and `GenError`'s together.
 dropGen :: MonadGenError m => GenT GE a -> GenT m a
 dropGen y = do
   r <- inspect y
@@ -373,19 +447,6 @@ dropGen y = do
     FatalError es -> fatalErrors es
     GenError es -> genErrors es
     Result a -> pure a
-
--- | Run one of the actions with frequency proportional to the count. If it fails, run the other.
-frequency2 :: forall m a. MonadGenError m => (Int, GenT GE a) -> (Int, GenT GE a) -> GenT m a
-frequency2 (n, g1) (m, g2)
-  | n <= 0 && m <= 0 = fatalError $ "Non positive frequencies in frequency2 " ++ show (n, m)
-  | n <= 0 = dropGen g2
-  | m <= 0 = dropGen g1
-  | True = do
-      i <- pureGen $ choose (1, n + m)
-      ans <- if i <= n then firstGenT [g1, g2] else firstGenT [g2, g1]
-      case ans of
-        Left _ -> fatalError "Both branches of frequency2 fail"
-        Right x -> pure x
 
 -- ======================================
