+++ libs/constrained-generators/src/Constrained/Graph.hs
@@ -2,7 +2,23 @@
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TupleSections #-}
 
-module Constrained.Graph where
+-- | This module provides a dependency graph implementation.
+module Constrained.Graph (
+  Graph,
+  opGraph,
+  nodes,
+  deleteNode,
+  subtractGraph,
+  dependency,
+  findCycle,
+  dependsOn,
+  dependencies,
+  noDependencies,
+  topsort,
+  transitiveClosure,
+  transitiveDependencies,
+  irreflexiveDependencyOn,
+) where
 
 import Control.Monad
 import Data.Foldable
@@ -14,6 +30,7 @@ import Data.Set (Set)
 import Data.Set qualified as Set
 import Prettyprinter
 
+-- | A graph with unlabeled edges for keeping track of dependencies
 data Graph node = Graph
   { edges :: !(Map node (Set node))
   , opEdges :: !(Map node (Set node))
@@ -38,12 +55,21 @@ instance Pretty n => Pretty (Graph n) where
         | (n, ns) <- Map.toList (edges gr)
         ]
 
+-- | Get all the nodes of a graph
 nodes :: Graph node -> Set node
 nodes (Graph e _) = Map.keysSet e
 
+-- | Delete a node from a graph
+deleteNode :: Ord node => node -> Graph node -> Graph node
+deleteNode x (Graph e o) = Graph (clean e) (clean o)
+  where
+    clean = Map.delete x . fmap (Set.delete x)
+
+-- | Invert the graph
 opGraph :: Graph node -> Graph node
 opGraph (Graph e o) = Graph o e
 
+-- | @subtractGraph g g'@ is the graph @g@ without the dependencies in @g'@
 subtractGraph :: Ord node => Graph node -> Graph node -> Graph node
 subtractGraph (Graph e o) (Graph e' o') =
   Graph
@@ -52,6 +78,8 @@ subtractGraph (Graph e o) (Graph e' o') =
   where
     del x y = Just $ Set.difference x y
 
+-- | @dependency x xs@ is the graph where @x@ depends on every node in @xs@
+-- and there are no other dependencies.
 dependency :: Ord node => node -> Set node -> Graph node
 dependency x xs =
   Graph
@@ -62,6 +90,7 @@ dependency x xs =
         (Map.fromList [(y, Set.singleton x) | y <- Set.toList xs])
     )
 
+-- | Every node in the first set depends on every node in the second set except themselves
 irreflexiveDependencyOn :: Ord node => Set node -> Set node -> Graph node
 irreflexiveDependencyOn xs ys =
   deps <> noDependencies ys
@@ -71,6 +100,7 @@ irreflexiveDependencyOn xs ys =
         (Map.fromDistinctAscList [(x, Set.delete x ys) | x <- Set.toList xs])
         (Map.fromDistinctAscList [(a, Set.delete a xs) | a <- Set.toList ys])
 
+-- | Get all down-stream dependencies of a node
 transitiveDependencies :: Ord node => node -> Graph node -> Set node
 transitiveDependencies x (Graph e _) = go (Set.singleton x) x
   where
@@ -78,16 +108,18 @@ transitiveDependencies x (Graph e _) = go (Set.singleton x) x
       where
         ys = fromMaybe mempty (Map.lookup y e)
 
+-- | Take the transitive closure of the graph
 transitiveClosure :: Ord node => Graph node -> Graph node
 transitiveClosure g = foldMap (\x -> dependency x (transitiveDependencies x g)) (nodes g)
 
+-- | The discrete graph containing all the input nodes without any dependencies
 noDependencies :: Ord node => Set node -> Graph node
 noDependencies ns = Graph nodeMap nodeMap
   where
     nodeMap = Map.fromList ((,mempty) <$> Set.toList ns)
 
--- | Topsort the graph, returning a cycle
--- (`Left cycle`) on failure.
+-- | Topsort the graph, returning either @Right order@ if the graph is a DAG or
+-- @Left cycle@  if it is not
 topsort :: Ord node => Graph node -> Either [node] [node]
 topsort gr@(Graph e _) = go [] e
   where
@@ -113,14 +145,11 @@ findCycle (Graph e _) node = concat . take 1 $ go mempty node
       where
         neighbours = maybe [] Set.toList $ Map.lookup n e
 
+-- | Get the dependencies of a node in the graph, `mempty` if the node is not
+-- in the graph
 dependencies :: Ord node => node -> Graph node -> Set node
 dependencies x (Graph e _) = fromMaybe mempty (Map.lookup x e)
 
+-- | Check if a node depends on another in the graph
 dependsOn :: Ord node => node -> node -> Graph node -> Bool
 dependsOn x y g = y `Set.member` dependencies x g
-
-deleteNode :: Ord node => node -> Graph node -> Graph node
-deleteNode x (Graph e o) =
-  Graph
-    (Map.delete x $ fmap (Set.delete x) e)
-    (Map.delete x $ fmap (Set.delete x) o)
