+++ libs/constrained-generators/src/Constrained/Properties.hs
@@ -26,6 +26,7 @@ conformsToSpecProp a s = case conformsToSpecE a (simplifySpec s) (pure "call to 
 forAllSpec :: (HasSpec a, QC.Testable p) => Specification a -> (a -> p) -> QC.Property
 forAllSpec spec prop = forAllSpecShow spec show prop
 
+-- | Like `forAllSpec` with a custom way of printing values
 forAllSpecShow ::
   (HasSpec a, QC.Testable p) => Specification a -> (a -> String) -> (a -> p) -> QC.Property
 forAllSpecShow spec pp prop =
