+++ libs/constrained-generators/src/Constrained/Spec/SumProd.hs
@@ -15,6 +15,7 @@
 {-# OPTIONS_GHC -Wno-orphans #-}
 {-# OPTIONS_GHC -Wno-redundant-constraints #-}
 
+-- | A lot of the surface-syntax related to generics
 module Constrained.Spec.SumProd (
   IsNormalType,
   ProdAsListComputes,
@@ -39,83 +40,22 @@ module Constrained.Spec.SumProd (
   nothing_,
   fst_,
   snd_,
-  fstW,
-  sndW,
   pair_,
-  injLeft_,
-  injRight_,
   prodFst_,
   prodSnd_,
   prod_,
-  PairSpec (..),
-  SumW (..),
 ) where
 
 import Constrained.AbstractSyntax
-import Constrained.Base (
-  BaseW (..),
-  Binder,
-  Forallable (..),
-  Fun (..),
-  GenericRequires,
-  HasSpec (..),
-  IsPred (..),
-  Pred,
-  Specification,
-  Term,
-  appTerm,
-  bind,
-  constrained,
-  fromGeneric_,
-  toGeneric_,
- )
-import Constrained.Conformance (
-  conformsToSpec,
-  satisfies,
- )
-import Constrained.Core (
-  Evidence (..),
- )
+import Constrained.Base
+import Constrained.Conformance
+import Constrained.Core
 import Constrained.Generation
-import Constrained.Generic (
-  ConstrOf,
-  HasSimpleRep (..),
-  Prod,
-  ProdOver,
-  SOP,
-  SimpleRep,
-  Sum,
-  SumOver,
-  TheSop,
-  (:::),
- )
-import Constrained.List (
-  All,
-  FunTy,
-  List (..),
-  MapList,
-  TypeList,
-  curryList,
-  listShape,
-  mapListC,
- )
-import Constrained.Syntax (
-  exists,
-  forAll,
-  letBind,
-  mkCase,
-  reify,
- )
-import Constrained.TheKnot (
-  FoldSpec (..),
-  FunW (..),
-  PairSpec (..),
-  ProdW (..),
-  preMapFoldSpec,
-  prodFst_,
-  prodSnd_,
-  prod_,
- )
+import Constrained.Generic
+import Constrained.List
+import Constrained.Spec.List
+import Constrained.Syntax
+import Constrained.TheKnot
 import Constrained.TypeErrors
 import Data.Typeable (Typeable)
 import GHC.Generics
@@ -123,93 +63,99 @@ import GHC.TypeLits (Symbol)
 import GHC.TypeNats
 import Test.QuickCheck (Arbitrary (..), oneof)
 
--- ==================================================================
--- Generics
--- HasSpec for various types that are Sums of Products
--- ==================================================================
+------------------------------------------------------------------------
+-- Syntax for `(,)` and `Either`
+------------------------------------------------------------------------
 
-instance (Typeable a, Typeable b) => HasSimpleRep (a, b)
+-- | `fst` in `Term` form
+fst_ :: (HasSpec x, HasSpec y) => Term (x, y) -> Term x
+fst_ = prodFst_ . toGeneric_
 
-instance (Typeable a, Typeable b, Typeable c) => HasSimpleRep (a, b, c)
+-- | `snd` in `Term` form
+snd_ :: (HasSpec x, HasSpec y) => Term (x, y) -> Term y
+snd_ = prodSnd_ . toGeneric_
 
-instance (Typeable a, Typeable b, Typeable c, Typeable d) => HasSimpleRep (a, b, c, d)
-
-instance (Typeable a, Typeable b, Typeable c, Typeable d, Typeable e) => HasSimpleRep (a, b, c, d, e)
-
-instance
-  (Typeable a, Typeable b, Typeable c, Typeable d, Typeable e, Typeable g) =>
-  HasSimpleRep (a, b, c, d, e, g)
-
-instance
-  (Typeable a, Typeable b, Typeable c, Typeable d, Typeable e, Typeable g, Typeable h) =>
-  HasSimpleRep (a, b, c, d, e, g, h)
-
-instance Typeable a => HasSimpleRep (Maybe a)
-
-instance (Typeable a, Typeable b) => HasSimpleRep (Either a b)
-
-instance
+-- | `(,)` in `Term` form
+pair_ ::
   ( HasSpec a
   , HasSpec b
-  ) =>
-  HasSpec (a, b)
-
-instance
-  ( HasSpec a
-  , HasSpec b
-  , HasSpec c
-  ) =>
-  HasSpec (a, b, c)
-
-instance
-  ( HasSpec a
-  , HasSpec b
-  , HasSpec c
-  , HasSpec d
-  ) =>
-  HasSpec (a, b, c, d)
-
-instance
-  ( HasSpec a
-  , HasSpec b
-  , HasSpec c
-  , HasSpec d
-  , HasSpec e
-  ) =>
-  HasSpec (a, b, c, d, e)
-
-instance
-  ( HasSpec a
-  , HasSpec b
-  , HasSpec c
-  , HasSpec d
-  , HasSpec e
-  , HasSpec g
-  ) =>
-  HasSpec (a, b, c, d, e, g)
-
-instance
-  ( HasSpec a
-  , HasSpec b
-  , HasSpec c
-  , HasSpec d
-  , HasSpec e
-  , HasSpec g
-  , HasSpec h
-  ) =>
-  HasSpec (a, b, c, d, e, g, h)
-
-instance
-  (IsNormalType a, HasSpec a) =>
-  HasSpec (Maybe a)
-
-instance
-  ( HasSpec a
   , IsNormalType a
-  , HasSpec b
   , IsNormalType b
   ) =>
-  HasSpec (Either a b)
+  Term a ->
+  Term b ->
+  Term (a, b)
+pair_ x y = fromGeneric_ $ prod_ x y
+
+-- | `Left` in `Term` form
+left_ ::
+  ( HasSpec a
+  , HasSpec b
+  , IsNormalType a
+  , IsNormalType b
+  ) =>
+  Term a ->
+  Term (Either a b)
+left_ = fromGeneric_ . injLeft_
+
+-- | `Right` in `Term` form
+right_ ::
+  ( HasSpec a
+  , HasSpec b
+  , IsNormalType a
+  , IsNormalType b
+  ) =>
+  Term b ->
+  Term (Either a b)
+right_ = fromGeneric_ . injRight_
+
+-- | @case .. of@ for `Term` and `Pred`. Note that the arguments
+-- here are @`Weighted` `Binder`@ over all the `Cases` of the
+-- `SimpleRep` of the scrutinee. The `Binder`s can be constructed with
+-- `branch` and `branchW`.
+caseOn ::
+  forall a.
+  ( GenericRequires a
+  , SimpleRep a ~ SumOver (Cases (SimpleRep a))
+  , TypeList (Cases (SimpleRep a))
+  ) =>
+  Term a ->
+  FunTy (MapList (Weighted Binder) (Cases (SimpleRep a))) Pred
+caseOn tm = curryList @(Cases (SimpleRep a)) (mkCase (toGeneric_ tm))
+
+-- | Build a branch in a `caseOn`
+branch ::
+  forall p a.
+  ( HasSpec a
+  , All HasSpec (Args a)
+  , IsPred p
+  , IsProd a
+  ) =>
+  FunTy (MapList Term (Args a)) p ->
+  Weighted Binder a
+branch body =
+  -- NOTE: It's not sufficient to simply apply `body` to all the arguments
+  -- with `uncurryList` because that will mean that `var` is repeated in the
+  -- body. For example, consider `branch $ \ i j -> i <=. j`. If we don't
+  -- build the lets this will boil down to `p :-> fst p <=. snd p` which
+  -- will blow up at generation time. If we instead do: `p :-> Let x (fst p) (Let y (snd p) (x <=. y))`
+  -- the solver will solve `x` and `y` separately (`y` before `x` in this case) and things
+  -- will work just fine.
+  Weighted Nothing (bind (buildBranch @p body . toArgs @a))
+
+-- | Build a branch in a `caseOn` with a weight attached.
+branchW ::
+  forall p a.
+  ( HasSpec a
+  , All HasSpec (Args a)
+  , IsPred p
+  , IsProd a
+  ) =>
+  Int ->
+  FunTy (MapList Term (Args a)) p ->
+  Weighted Binder a
+branchW w body =
+  Weighted (Just w) (bind (buildBranch @p body . toArgs @a))
 
 -- ====================================================
 -- All the magic for things like 'caseOn', 'match', forAll' etc. lives here.
@@ -235,6 +181,7 @@ type family ResultType t where
   ResultType (a -> b) = ResultType b
   ResultType a = a
 
+-- | A normal type, not an underlying generic representation using `Sum` and t`Prod`
 type IsNormalType a =
   ( AssertComputes
       (Cases a)
@@ -258,6 +205,7 @@ type family Cases t where
   Cases (Sum a b) = a : Cases b
   Cases a = '[a]
 
+-- | A single-constructor type like t`(,)`
 type IsProductType a =
   ( HasSimpleRep a
   , AssertComputes
@@ -366,15 +314,7 @@ instance
   where
   mkCases r k = Weighted Nothing (bind (r @(ProdOver as))) :> mkCases @c @_ @cs r k
 
-------------------------------------------------------------------------
--- Syntax
-------------------------------------------------------------------------
-
-fst_ :: (HasSpec x, HasSpec y) => Term (x, y) -> Term x
-fst_ = appTerm (ComposeW ProdFstW ToGenericW)
-
-snd_ :: (HasSpec x, HasSpec y) => Term (x, y) -> Term y
-snd_ = appTerm (ComposeW ProdSndW ToGenericW)
+-- Instances --------------------------------------------------------------
 
 fstW :: (HasSpec a, HasSpec b) => FunW '[(a, b)] a
 fstW = ComposeW ProdFstW ToGenericW
@@ -395,47 +335,6 @@ instance
   shrink NoFold = []
   shrink FoldSpec {} = [NoFold]
 
-pair_ ::
-  ( HasSpec a
-  , HasSpec b
-  , IsNormalType a
-  , IsNormalType b
-  ) =>
-  Term a ->
-  Term b ->
-  Term (a, b)
-pair_ x y = fromGeneric_ $ prod_ x y
-
-left_ ::
-  ( HasSpec a
-  , HasSpec b
-  , IsNormalType a
-  , IsNormalType b
-  ) =>
-  Term a ->
-  Term (Either a b)
-left_ = fromGeneric_ . injLeft_
-
-right_ ::
-  ( HasSpec a
-  , HasSpec b
-  , IsNormalType a
-  , IsNormalType b
-  ) =>
-  Term b ->
-  Term (Either a b)
-right_ = fromGeneric_ . injRight_
-
-caseOn ::
-  forall a.
-  ( GenericRequires a
-  , SimpleRep a ~ SumOver (Cases (SimpleRep a))
-  , TypeList (Cases (SimpleRep a))
-  ) =>
-  Term a ->
-  FunTy (MapList (Weighted Binder) (Cases (SimpleRep a))) Pred
-caseOn tm = curryList @(Cases (SimpleRep a)) (mkCase (toGeneric_ tm))
-
 buildBranch ::
   forall p as.
   ( All HasSpec as
@@ -447,38 +346,6 @@ buildBranch ::
 buildBranch bd Nil = toPred bd
 buildBranch bd (t :> args) =
   letBind t $ \x -> buildBranch @p (bd x) args
-
-branch ::
-  forall p a.
-  ( HasSpec a
-  , All HasSpec (Args a)
-  , IsPred p
-  , IsProd a
-  ) =>
-  FunTy (MapList Term (Args a)) p ->
-  Weighted Binder a
-branch body =
-  -- NOTE: It's not sufficient to simply apply `body` to all the arguments
-  -- with `uncurryList` because that will mean that `var` is repeated in the
-  -- body. For example, consider `branch $ \ i j -> i <=. j`. If we don't
-  -- build the lets this will boil down to `p :-> fst p <=. snd p` which
-  -- will blow up at generation time. If we instead do: `p :-> Let x (fst p) (Let y (snd p) (x <=. y))`
-  -- the solver will solve `x` and `y` separately (`y` before `x` in this case) and things
-  -- will work just fine.
-  Weighted Nothing (bind (buildBranch @p body . toArgs @a))
-
-branchW ::
-  forall p a.
-  ( HasSpec a
-  , All HasSpec (Args a)
-  , IsPred p
-  , IsProd a
-  ) =>
-  Int ->
-  FunTy (MapList Term (Args a)) p ->
-  Weighted Binder a
-branchW w body =
-  Weighted (Just w) (bind (buildBranch @p body . toArgs @a))
 
 -- | ProdAsListComputes is here to make sure that in situations like this:
 --
@@ -511,6 +378,7 @@ type ProdAsListComputes a =
     (Text "Have you considered adding an IsNormalType or ProdAsListComputes constraint?")
     (ProductAsList a)
 
+-- | Pattern-match on a product type and build constraints with the constituents:
 match ::
   forall p a.
   ( IsProductType a
@@ -523,6 +391,10 @@ match p m = caseOn p (branch @p m)
 
 -- NOTE: `ResultType r ~ Term a` is NOT a redundant constraint,
 -- removing it causes type inference to break elsewhere
+
+-- | Create a constructor @c@:
+-- > just_ :: (HasSpec a, IsNormalType a) => Term a -> Term (Maybe a)
+-- > just_ = con @"Just"
 con ::
   forall c a r.
   ( SimpleRep a ~ SOP (TheSop a)
@@ -539,12 +411,20 @@ con =
   curryList @(ConstrOf c (TheSop a)) @Term
     (fromGeneric_ @a . inj_ @c @(TheSop a) . prodOver_)
 
+-- | `Term`-level `Just`
 just_ :: (HasSpec a, IsNormalType a) => Term a -> Term (Maybe a)
 just_ = con @"Just"
 
+-- | `Term`-level `Nothing`
 nothing_ :: (HasSpec a, IsNormalType a) => Term (Maybe a)
 nothing_ = con @"Nothing" (Lit ())
 
+-- | Select a specific field from a single-constructor type:
+-- > data Record = Record { foo :: Int, bar :: Bool }
+-- > foo_ :: Term Record -> Term Int
+-- > foo_ = sel @0
+-- > bar_ :: Term Record -> Term Bool
+-- > bar_ = sel @1
 sel ::
   forall n a c as.
   ( SimpleRep a ~ ProdOver as
@@ -636,6 +516,8 @@ instance
   prodOver_ (a :> as) = prod_ a (prodOver_ as)
 
 -- TODO: the constraints around this are horrible!! We should figure out a way to make these things nicer.
+
+-- | `caseOn` a _single_ constructor only
 onCon ::
   forall c a p.
   ( IsConstrOf c (ProdOver (ConstrOf c (TheSop a))) (TheSop a)
@@ -659,6 +541,9 @@ onCon tm p =
         (buildBranch @p p . toArgs)
     )
 
+-- | Check if a value is an instance of a specific constructor:
+-- > isJustConstraint :: HasSpec a => Term (Maybe a) -> Pred
+-- > isJustConstraint t = isCon @"Just" t
 isCon ::
   forall c a.
   ( IsConstrOf c (ProdOver (ConstrOf c (TheSop a))) (TheSop a)
@@ -677,6 +562,7 @@ isCon tm =
         (const $ Assert (Lit True))
     )
 
+-- | `onCon` specialized to `Just`
 onJust ::
   forall a p.
   (HasSpec a, IsNormalType a, IsPred p) =>
@@ -685,6 +571,7 @@ onJust ::
   Pred
 onJust = onCon @"Just"
 
+-- | `isCon` specialized to `Just`
 isJust ::
   forall a.
   (HasSpec a, IsNormalType a) =>
@@ -713,3 +600,90 @@ data Picky = PickFirst | PickSecond deriving (Ord, Eq, Show, Generic)
 instance HasSimpleRep Picky
 
 instance HasSpec Picky
+
+------------------------------------------------------------------------
+-- Some generic instances of HasSpec and HasSimpleRep
+------------------------------------------------------------------------
+
+instance (Typeable a, Typeable b) => HasSimpleRep (a, b)
+
+instance (Typeable a, Typeable b, Typeable c) => HasSimpleRep (a, b, c)
+
+instance (Typeable a, Typeable b, Typeable c, Typeable d) => HasSimpleRep (a, b, c, d)
+
+instance (Typeable a, Typeable b, Typeable c, Typeable d, Typeable e) => HasSimpleRep (a, b, c, d, e)
+
+instance
+  (Typeable a, Typeable b, Typeable c, Typeable d, Typeable e, Typeable g) =>
+  HasSimpleRep (a, b, c, d, e, g)
+
+instance
+  (Typeable a, Typeable b, Typeable c, Typeable d, Typeable e, Typeable g, Typeable h) =>
+  HasSimpleRep (a, b, c, d, e, g, h)
+
+instance Typeable a => HasSimpleRep (Maybe a)
+
+instance (Typeable a, Typeable b) => HasSimpleRep (Either a b)
+
+instance
+  ( HasSpec a
+  , HasSpec b
+  ) =>
+  HasSpec (a, b)
+
+instance
+  ( HasSpec a
+  , HasSpec b
+  , HasSpec c
+  ) =>
+  HasSpec (a, b, c)
+
+instance
+  ( HasSpec a
+  , HasSpec b
+  , HasSpec c
+  , HasSpec d
+  ) =>
+  HasSpec (a, b, c, d)
+
+instance
+  ( HasSpec a
+  , HasSpec b
+  , HasSpec c
+  , HasSpec d
+  , HasSpec e
+  ) =>
+  HasSpec (a, b, c, d, e)
+
+instance
+  ( HasSpec a
+  , HasSpec b
+  , HasSpec c
+  , HasSpec d
+  , HasSpec e
+  , HasSpec g
+  ) =>
+  HasSpec (a, b, c, d, e, g)
+
+instance
+  ( HasSpec a
+  , HasSpec b
+  , HasSpec c
+  , HasSpec d
+  , HasSpec e
+  , HasSpec g
+  , HasSpec h
+  ) =>
+  HasSpec (a, b, c, d, e, g, h)
+
+instance
+  (IsNormalType a, HasSpec a) =>
+  HasSpec (Maybe a)
+
+instance
+  ( HasSpec a
+  , IsNormalType a
+  , HasSpec b
+  , IsNormalType b
+  ) =>
+  HasSpec (Either a b)
